Fix dxPolarChart with scale breaks
